fix: MyPlantController 내 registerPlant 메서드 plantId 파라미터 데이터 타입 수정

+++ src/main/java/com/cookeep/cookeep/api/controller/MyPlantController.java
@@ -40,7 +40,7 @@ public class MyPlantController {
     })
     @PostMapping("/{plantId}") // /api/my-plants/{plantId}
     public DataResponse<Void> registerPlant(
-            @Parameter(description = "기본 식물 ID") @PathVariable int plantId) {
+            @Parameter(description = "기본 식물 ID") @PathVariable long plantId) {
         Long userId = userProvider.getCurrentUserId();
         userPlantService.registerPlant(userId, plantId);
         return DataResponse.from(null);
